Refactor to BitOrder enum for ACHistory
